spring introduced to project, will create separate branch for api test
